Drop stdio.h include
The sha1.c file includes stdio.h without seemingly using it for anyhing.
It doesn't make sense for a hash library to use stdio and having that
header included makes it harder to use the library in freestanding
environment.

Drop the include to make it easier to use the library in freestanding
projects.